Try to use the most precise type (int to tinyint)
"For optimum storage, you should try to use the most precise type in all
cases."
https://dev.mysql.com/doc/refman/5.7/en/choosing-types.html

awards
[prefix]_awards
Possible values: 1 or 2
`typ` TINYINT(1) NOT NULL,

events
[prefix]_events
Possible values: 0 or 1
`show` TINYINT(1) NOT NULL,
[prefix]_events_entrants
Possible values: 1 or 2
`status` TINYINT(1) NOT NULL

forum
[prefix]_forum_items
Possible values: 0 or 1
`type` TINYINT(1) NOT NULL,
[prefix]_forum_topics
`type` TINYINT(1) NOT NULL DEFAULT 0,
Possible values: 0 or 1
`status` TINYINT(1) NOT NULL DEFAULT 0,
Possible values: 0 or 1

gallery
[prefix]_gallery_items
Possible values: 0 or 1
`type` TINYINT(1) NOT NULL,

guestbook
[prefix]_gbook
Possible values: 0 or 1
`setfree` TINYINT(1) NOT NULL DEFAULT 0,

jobs
[prefix]_jobs
Possible values: 0 or 1
`show` TINYINT(1) NOT NULL,

admin
[prefix]_menu_items
Possible values: 0, 1, 2, 3 or 4
`type` TINYINT(1) NOT NULL,

partner
[prefix]_partners
Possible values: 0 or 1
`setfree` TINYINT(11) NOT NULL DEFAULT 0,

privacy
[prefix]_privacy
Possible values: 0 or 1
`show` TINYINT(1) NOT NULL,

user
[prefix]_users
Possible values: 0 or 1
`opt_mail` TINYINT(1) DEFAULT 1,
Possible values: 0 or 1
`opt_gallery` TINYINT(1) DEFAULT 1,
Possible values: 0 or 1
`confirmed` TINYINT(1) DEFAULT 1,
[prefix]_profile_fields
Possible values: 0 or 1
`type` TINYINT(1) NOT NULL,
[prefix]_users_dialog_reply
Possible values: 0 or 1
`read` TINYINT(1) DEFAULT 0,
[prefix]_users_gallery_items
Possible values: 0 or 1
`type` TINYINT(1) NOT NULL,

+++ application/modules/admin/config/config.php
@@ -116,7 +116,7 @@ class Config extends \Ilch\Config\Install
                   `page_id` INT(11) NOT NULL DEFAULT 0,
                   `box_id` INT(11) NOT NULL DEFAULT 0,
                   `box_key` VARCHAR(255) NULL DEFAULT NULL,
-                  `type` INT(11) NOT NULL,
+                  `type` TINYINT(1) NOT NULL,
                   `title` VARCHAR(255) NOT NULL,
                   `href` VARCHAR(255) NULL DEFAULT NULL,
                   `module_key` VARCHAR(255) NULL DEFAULT NULL,
