+++ application/modules/privacy/config/config.php
@@ -37,7 +37,7 @@ class Config extends \Ilch\Config\Install
                   `urltitle` VARCHAR(255) NOT NULL,
                   `url` VARCHAR(255) NOT NULL,
                   `text` MEDIUMTEXT NOT NULL,
-                  `show` INT(11) NOT NULL,
+                  `show` TINYINT(1) NOT NULL,
                   PRIMARY KEY (`id`)
                 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci AUTO_INCREMENT=1;
